renamed component name
